Don't use bool type for obf
The type was getting confused since we were converting from a void* to a bool.
I assumed the truth value would look at the entire byte (i.e., the cast would
test the truth value of the byte), but no such luck. Going back to an
old-style int for this "flag" makes things work. For example, the assignment
in the consume function now zeros an entire word.